Source-Code/Backend/mqtt-controller/application/requirements.txt
flask
kubernetes
mongoengine
pika
protobuf
redis
uuid
waitress